backend/requirements.txt
asgiref==3.5.0
click==8.1.2
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
joblib==1.1.0
nltk==3.7
numpy==1.22.3
pytz==2022.1
regex==2022.3.15
sqlparse==0.4.2
tqdm==4.64.0
